[SYCL][NFC] Group all msvc specific math functions together
Move double and float version of msvc specific functions in single
file so that we won't forget to fix double version bug.
